Don't attempt to add version control info for now
This is a new Go 1.18 feature. With this change, the behaviour is as
earlier versions of Go.

Without this change, the build fails because Git is not installed, and
the default Go build behaviour when the `.git` directory is present is
to try to run `git` to embed this information.